feat: 🎸 commonize the relations
