--- app/models/concerns/has_chapters.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module HasChapters
-  extend ActiveSupport::Concern
-
-  def chapters
-    @chapters ||= EpisodeChapter.where(episode_number: number).all
-  end
-end
