+++ app/models/episode/has_guest_interviews.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class Episode
+  module HasGuestInterviews
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :guest_interviews, foreign_key: :episode_number, primary_key: :number, inverse_of: :episode
+      has_many :guest_interview_profiles, through: :guest_interviews
+      has_many :guests, through: :guest_interviews
+    end
+  end
+end
