+++ app/models/episode/has_references.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class Episode
+  module HasReferences
+    extend ActiveSupport::Concern
+
+    included do
+      has_many :references, class_name: 'EpisodeReference', foreign_key: :episode_number, primary_key: :number,
+                            inverse_of: :episode
+    end
+  end
+end
